Added form for managers to create a restaurant menu
